simulator/requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0.1
influxdb-client==1.36.1
